requirements.txt
streamlit
pandas
scikit-learn
groq
python-dotenv
xgboost